refactor(run): extract compaction policy into run/compaction_policy.rs (Wave C1)
First cut of the run.rs split (3,955 -> 3,613 lines). Moves the pure
compaction/iteration policy — DEFAULT_COMPACT_THRESHOLD, compact_threshold,
resolve_compact_threshold, soft_compact_threshold, HARD/SOFT_COMPACT_PRESERVE,
CompactionTier, CompactionOutcome, DEFAULT_MAX_ITERATIONS, max_iterations,
pick_compact_plan — plus its 15 unit tests into a new run/ submodule. The
runtime-mutating maybe_compact_session (and its 2 runtime-building tests) stay
in run.rs as the policy's caller.

Behavior-identical pure code movement (no A/B needed): every external run::X
path is preserved via re-export (pub use for compact_threshold/
soft_compact_threshold/max_iterations/DEFAULT_*; pub(crate) use for
pick_compact_plan/CompactionOutcome). CompactionTier isn't re-exported (used
only by run.rs's own test, which now imports it from the submodule).

whole-crate cargo build (re-resolves every run:: caller) green; fmt + clippy
--all-targets (lean + --all-features) -D warnings clean; cargo test --lib 1035
passed (count unchanged — tests moved, not dropped).

Next in Wave C: C2 recall_index, C3 cli_prompter, C4 runtime_build, C5 forge
(Claude-led), C6 repl, C-final (drop the crate-wide too_many_lines/
too_many_arguments allows). Map in codev_plan_waves_C_D.
